pyproject.toml
# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=config.settings.test --reuse-db --import-mode=importlib"
python_files = [
    "tests.py",
    "test_*.py",
]

# ==== Coverage ====
[tool.coverage.run]
include = ["erp/**"]
omit = ["*/migrations/*", "*/tests/*"]
plugins = ["django_coverage_plugin"]

# ==== mypy ====
[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.test"

# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/djlint/djLint/issues/687
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2

[tool.ruff]
target-version = "py312"
# Exclude a variety of commonly ignored directories.
extend-exclude = [
    "*/migrations/*.py",
    "staticfiles/*",
]

[tool.ruff.lint]
select = [
  "F",
  "E",
  "W",
  "C90",
  "I",
  "N",
  "UP",
  "YTT",
  # "ANN", # flake8-annotations: we should support this in the future but 100+ errors atm
  "ASYNC",
  "S",
  "BLE",
  "FBT",
  "B",
  "A",
  "COM",
  "C4",
  "DTZ",
  "T10",
  "DJ",
  "EM",
  "EXE",
  "FA",
  'ISC',
  "ICN",
  "G",
  'INP',
  'PIE',
  "T20",
  'PYI',
  'PT',
  "Q",
  "RSE",
  "RET",
  "SLF",
  "SLOT",
  "SIM",
  "TID",
  "TC",
  "INT",
  # "ARG", # Unused function argument
  "PTH",
  "ERA",
  "PD",
  "PGH",
  "PL",
  "TRY",
  "FLY",
  # "NPY",
  # "AIR",
  "PERF",
  # "FURB",
  # "LOG",
  "RUF",
]
ignore = [
  "S101", # Use of assert detected https://docs.astral.sh/ruff/rules/assert/
  "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
  "SIM102", # sometimes it's better to nest
  "UP038", # Checks for uses of isinstance/issubclass that take a tuple
          # of types for comparison.
          # Deactivated because it can make the code slow:
          # https://github.com/astral-sh/ruff/issues/7871
]
# The fixes in extend-unsafe-fixes will require
# provide the `--unsafe-fixes` flag when fixing.
extend-unsafe-fixes = [
  "UP038",
]

[tool.ruff.lint.isort]
force-single-line = true


[tool.poetry.group.dev.dependencies]
werkzeug = {version = "3.1.3", extras = ["watchdog"]}
ipdb = "0.13.13"
psycopg = {version = "3.2.5", extras = ["binary"]}
watchfiles = "1.0.4"
mypy = "1.15.0"
django-stubs = {version = "5.1.3", extras = ["compatible-mypy"]}
pytest = "8.3.4"
pytest-sugar = "1.0.0"
djangorestframework-stubs = "3.15.3"
sphinx = "8.2.1"
sphinx-autobuild = "2024.10.3"
ruff = "0.9.7"
coverage = "7.6.12"
djlint = "1.36.4"
pre-commit = "4.1.0"
factory-boy = "3.3.2"
django-debug-toolbar = "5.0.1"
django-extensions = "3.2.3"
django-coverage-plugin = "3.1.0"
pytest-django = "4.10.0"

[project]
name = "erp"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "python-slugify (==8.0.4)",
    "pillow (==11.1.0)",
    "argon2-cffi (==23.1.0)",
    "whitenoise (==6.9.0)",
    "redis (==5.2.1)",
    "celery (==5.4.0)",
    "django (==5.0.12)",
    "django-model-utils (==5.0.0)",
    "django-crispy-forms (==2.3)",
    "crispy-bootstrap5 (==2024.10)",
    "django-redis (==5.4.0)",
    "djangorestframework (==3.15.2)",
    "django-celery-beat (==2.7.0)",
    "django-environ (==0.12.0)",
    "django-allauth[mfa] (==65.4.1)",
    "django-cors-headers (==4.7.0)",
    "drf-spectacular (==0.28.0)",
    "djoser (>=2.3.1,<3.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "django-phonenumber-field (>=8.0.0,<9.0.0)",
    "phonenumbers (>=8.13.55,<9.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-widget-tweaks (>=1.5.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
